refactor(dto): cleanse and format all dtos
