--- src/main/java/wooteco/subway/dto/LineBasicResponse.java
@@ -1,46 +1,0 @@
-package wooteco.subway.dto;
-
-import java.util.Objects;
-
-public class LineBasicResponse {
-
-    private final Long id;
-    private final String name;
-    private final String color;
-
-    public LineBasicResponse(Long id, String name, String color) {
-        this.id = id;
-        this.name = name;
-        this.color = color;
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        LineBasicResponse that = (LineBasicResponse) o;
-        return Objects.equals(id, that.id) && Objects.equals(name, that.name)
-                && Objects.equals(color, that.color);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(id, name, color);
-    }
-}
