+++ src/main/java/wooteco/subway/dto/StationRequest.java
@@ -1,6 +1,7 @@
 package wooteco.subway.dto;
 
 public class StationRequest {
+    
     private String name;
 
     public StationRequest() {
